Struct rustc_type_ir::binder::ArgFolder

struct ArgFolder<'a, I: Interner> {
    cx: I,
    args: &'a [I::GenericArg],
    binders_passed: u32,
}

Fields
cx: Iargs: &'a [I::GenericArg]binders_passed: u32

Number of region binders we have passed through while doing the instantiation

Implementations

impl<'a, I: Interner> ArgFolder<'a, I>

fn ty_for_param(&self, p: I::ParamTy, source_ty: I::Ty) -> I::Ty

fn type_param_expected( &self, p: I::ParamTy, ty: I::Ty, kind: GenericArgKind<I>, ) -> !

fn type_param_out_of_range(&self, p: I::ParamTy, ty: I::Ty) -> !

fn const_for_param(&self, p: I::ParamConst, source_ct: I::Const) -> I::Const

fn const_param_expected( &self, p: I::ParamConst, ct: I::Const, kind: GenericArgKind<I>, ) -> !

fn const_param_out_of_range(&self, p: I::ParamConst, ct: I::Const) -> !

fn region_param_expected( &self, ebr: I::EarlyParamRegion, r: I::Region, kind: GenericArgKind<I>, ) -> !

fn region_param_out_of_range(&self, ebr: I::EarlyParamRegion, r: I::Region) -> !

fn shift_vars_through_binders<T: TypeFoldable<I>>(&self, val: T) -> T

It is sometimes necessary to adjust the De Bruijn indices during instantiation. This occurs when we are instantating a type with escaping bound vars into a context where we have passed through binders. That’s quite a mouthful. Let’s see an example:

type Func<A> = fn(A);
type MetaFunc = for<'a> fn(Func<&'a i32>);

The type MetaFunc, when fully expanded, will be

for<'a> fn(fn(&'a i32))
//      ^~ ^~ ^~~
//      |  |  |
//      |  |  DebruijnIndex of 2
//      Binders

Here the 'a lifetime is bound in the outer function, but appears as an argument of the inner one. Therefore, that appearance will have a DebruijnIndex of 2, because we must skip over the inner binder (remember that we count De Bruijn indices from 1). However, in the definition of MetaFunc, the binder is not visible, so the type &'a i32 will have a De Bruijn index of 1. It’s only during the instantiation that we can see we must increase the depth by 1 to account for the binder that we passed through.

As a second example, consider this twist:

type FuncTuple<A> = (A,fn(A));
type MetaFuncTuple = for<'a> fn(FuncTuple<&'a i32>);

Here the final type will be:

for<'a> fn((&'a i32, fn(&'a i32)))
//          ^~~         ^~~
//          |           |
//   DebruijnIndex of 1 |
//               DebruijnIndex of 2

As indicated in the diagram, here the same type &'a i32 is instantiated once, but in the first case we do not increase the De Bruijn index and in the second case we do. The reason is that only in the second case have we passed through a fn binder.

fn shift_region_through_binders(&self, region: I::Region) -> I::Region