Trait rustc_type_ir::inherent::EvaluationCache

pub trait EvaluationCache<I: Interner> {
    // Required methods
    fn insert(
        &self,
        tcx: I,
        key: CanonicalInput<I>,
        proof_tree: Option<I::CanonicalGoalEvaluationStepRef>,
        additional_depth: usize,
        encountered_overflow: bool,
        cycle_participants: FxHashSet<CanonicalInput<I>>,
        dep_node: I::DepNodeIndex,
        result: QueryResult<I>,
    );
    fn get(
        &self,
        tcx: I,
        key: CanonicalInput<I>,
        stack_entries: impl IntoIterator<Item = CanonicalInput<I>>,
        available_depth: usize,
    ) -> Option<CacheData<I>>;
}

Required Methods

fn insert( &self, tcx: I, key: CanonicalInput<I>, proof_tree: Option<I::CanonicalGoalEvaluationStepRef>, additional_depth: usize, encountered_overflow: bool, cycle_participants: FxHashSet<CanonicalInput<I>>, dep_node: I::DepNodeIndex, result: QueryResult<I>, )

Insert a final result into the global cache.

fn get( &self, tcx: I, key: CanonicalInput<I>, stack_entries: impl IntoIterator<Item = CanonicalInput<I>>, available_depth: usize, ) -> Option<CacheData<I>>

Try to fetch a cached result, checking the recursion limit and handling root goals of coinductive cycles.

If this returns Some the cache result can be used.

Object Safety

This trait is not object safe.