rustc_type_ir::predicate
Struct TraitPredicate

pub struct TraitPredicate<I: Interner> {
    pub trait_ref: TraitRef<I>,
    pub polarity: PredicatePolarity,
}

Fields
trait_ref: TraitRef<I>polarity: PredicatePolarity

If polarity is Positive: we are proving that the trait is implemented.

If polarity is Negative: we are proving that a negative impl of this trait exists. (Note that coherence also checks whether negative impls of supertraits exist via a series of predicates.)

If polarity is Reserved: that’s a bug.

Implementations

impl<I: Interner> TraitPredicate<I>

pub fn with_self_ty(self, interner: I, self_ty: I::Ty) -> Self

pub fn def_id(self) -> I::DefId

pub fn self_ty(self) -> I::Ty