Struct rustc_type_ir::EarlyBinder

pub struct EarlyBinder<I: Interner, T> {
    value: T,
    _tcx: PhantomData<I>,
}

Similar to super::Binder except that it tracks early bound generics, i.e. struct Foo<T>(T) needs T instantiated immediately. This type primarily exists to avoid forgetting to call instantiate.

If you don’t have anything to instantiate, you may be looking for instantiate_identity or skip_binder.

Fields
value: T_tcx: PhantomData<I>
Implementations

impl<I: Interner, T> EarlyBinder<I, T>

pub fn bind(value: T) -> EarlyBinder<I, T>

pub fn as_ref(&self) -> EarlyBinder<I, &T>

pub fn map_bound_ref<F, U>(&self, f: F) -> EarlyBinder<I, U>
where F: FnOnce(&T) -> U,

pub fn map_bound<F, U>(self, f: F) -> EarlyBinder<I, U>
where F: FnOnce(T) -> U,

pub fn try_map_bound<F, U, E>(self, f: F) -> Result<EarlyBinder<I, U>, E>
where F: FnOnce(T) -> Result<U, E>,

pub fn rebind<U>(&self, value: U) -> EarlyBinder<I, U>

pub fn skip_binder(self) -> T

Skips the binder and returns the “bound” value. This can be used to extract data that does not depend on generic parameters (e.g., getting the DefId of the inner value or getting the number of arguments of an FnSig). Otherwise, consider using instantiate_identity.

To skip the binder on x: &EarlyBinder<I, T> to obtain &T, leverage EarlyBinder::as_ref: x.as_ref().skip_binder().

See also Binder::skip_binder, which is the analogous operation on super::Binder.

impl<I: Interner, T> EarlyBinder<I, Option<T>>

pub fn transpose(self) -> Option<EarlyBinder<I, T>>

impl<I: Interner, Iter: IntoIterator> EarlyBinder<I, Iter>
where Iter::Item: TypeFoldable<I>,

pub fn iter_instantiated<A>( self, cx: I, args: A, ) -> IterInstantiated<I, Iter, A>
where A: SliceLike<Item = I::GenericArg>,

pub fn iter_identity(self) -> Iter::IntoIter

Similar to instantiate_identity, but on an iterator of TypeFoldable values.

impl<'s, I: Interner, Iter: IntoIterator> EarlyBinder<I, Iter>
where Iter::Item: Deref, <<Iter as IntoIterator>::Item as Deref>::Target: Copy + TypeFoldable<I>,

pub fn iter_instantiated_copied( self, cx: I, args: &'s [I::GenericArg], ) -> IterInstantiatedCopied<'s, I, Iter>

pub fn iter_identity_copied( self, ) -> impl Iterator<Item = <<Iter as IntoIterator>::Item as Deref>::Target>

Similar to instantiate_identity, but on an iterator of values that deref to a TypeFoldable.

impl<I: Interner, T: IntoIterator> EarlyBinder<I, T>

pub fn transpose_iter(self) -> EarlyBinderIter<I, T::IntoIter>

impl<I: Interner, T: TypeFoldable<I>> EarlyBinder<I, T>

pub fn instantiate<A>(self, cx: I, args: A) -> T
where A: SliceLike<Item = I::GenericArg>,

pub fn instantiate_identity(self) -> T

Makes the identity replacement T0 => T0, ..., TN => TN. Conceptually, this converts universally bound variables into placeholders when inside of a given item.

For example, consider for<T> fn foo<T>(){ .. }:

Outside of foo, T is bound (represented by the presence of EarlyBinder).

Inside of the body of foo, we treat T as a placeholder by calling instantiate_identity to discharge the EarlyBinder.

pub fn no_bound_vars(self) -> Option<T>

Returns the inner value, but only if it contains no bound vars.