Module rustc_type_ir::sty

Enums

DynKind

Specifies how a trait object is represented.

RegionKind

Representation of regions. Note that the NLL checker uses a distinct representation of regions. For this reason, it internally replaces all the regions with inference variables – the index of the variable is then used to index into internal NLL data structures. See rustc_const_eval::borrow_check module for more information.

TyKind

Defines the kinds of types used by the type system.

Functions

regionkind_discriminant 🔒

tykind_discriminant 🔒