rustc_type_ir

Module ty_kind

Re-exports

pub use self::closure::*;

Modules

closure 🔒

Structs

AliasTy

Represents the projection of an associated, opaque, or lazy-type-alias type.

CoroutineWitnessTypes

FloatVid

A floating-point (f32 or f64) type variable ID.

FnHeader

FnSig

FnSigTys

IntVid

An integral (u32, i32, usize, etc.) type variable ID.

TyVid

A type variable ID.

TypeAndMut

UnsafeBinderInner

Enums

AliasTyKind

DynKind

Specifies how a trait object is represented.

FloatVarValue

InferTy

A placeholder for a type that hasn’t been inferred yet.

IntVarValue

TyKind

Defines the kinds of types used by the type system.