rustc_type_ir::ty_kind

Module closure

Structs

ClosureArgs

A closure can be modeled as a struct that looks like:

ClosureArgsParts

Struct returned by split().

CoroutineArgs

Similar to ClosureArgs; see the above documentation for more.

CoroutineArgsParts

CoroutineClosureArgs

CoroutineClosureArgsParts

See docs for explanation of how each argument is used.

CoroutineClosureSignature

FoldEscapingRegions 🔒

Instantiates a for<'env> ... binder with a specific region.

GenSig

HasRegionsBoundAt 🔒

Unlike has_escaping_bound_vars or outermost_exclusive_binder, this will detect only regions bound at the debruijn index.