rustc_type_ir::ty_kind::closure
Struct CoroutineArgsParts

pub struct CoroutineArgsParts<I: Interner> {
    pub parent_args: I::GenericArgsSlice,
    pub kind_ty: I::Ty,
    pub resume_ty: I::Ty,
    pub yield_ty: I::Ty,
    pub return_ty: I::Ty,
    pub witness: I::Ty,
    pub tupled_upvars_ty: I::Ty,
}

Fields
parent_args: I::GenericArgsSlice

This is the args of the typeck root.

kind_ty: I::Ty

The coroutines returned by a coroutine-closure’s AsyncFnOnce/AsyncFnMut implementations must be distinguished since the former takes the closure’s upvars by move, and the latter takes the closure’s upvars by ref.

This field distinguishes these fields so that codegen can select the right body for the coroutine. This has the same type representation as the closure kind: i8/i16/i32.

For regular coroutines, this field will always just be ().

resume_ty: I::Tyyield_ty: I::Tyreturn_ty: I::Tywitness: I::Ty

The interior type of the coroutine. Represents all types that are stored in locals in the coroutine’s body.

tupled_upvars_ty: I::Ty

The upvars captured by the closure. Remains an inference variable until the upvar analysis, which happens late in HIR typeck.