Crate rustc_type_ir

Re-exports

pub use codec::*;

pub use sty::*;

pub use ty_info::*;

Modules

codec

fold

A folding traversal mechanism for complex data structures that contain type information.

macros 🔒

structural_impls 🔒

This module contains implementations of the TypeFoldable and TypeVisitable traits for various types in the Rust compiler. Most are written by hand, though we’ve recently added some macros and proc-macros to help with the tedium.

sty

ty_info

visit

A visiting traversal mechanism for complex data structures that contain type information.

Structs

DebruijnIndex

A De Bruijn index is a standard means of representing regions (and perhaps later types) in a higher-ranked setting. In particular, imagine a type like this:

FloatVarValue

FloatVid

An floating-point (f32 or f64) type variable ID.

IntVid

An integral (u32, i32, usize, etc.) type variable ID.

OptWithInfcx

TyVid

A type variable ID.

TypeFlags

Flags that we track on types. These flags are propagated upwards through the type during type construction, so that we can quickly check whether the type has various kinds of types in it without recursing over the type itself.

UniverseIndex

“Universes” are used during type- and trait-checking in the presence of for<..> binders to control what sets of names are visible. Universes are arranged into a tree: the root universe contains names that are always visible. Each child then adds a new set of names that are visible, in addition to those of its parent. We say that the child universe “extends” the parent universe with new names.

Enums

FloatTy

InferTy

A placeholder for a type that hasn’t been inferred yet.

IntTy

IntVarValue

UintTy

Variance

Constants

INNERMOST

Traits

CollectAndApply

Imagine you have a function F: FnOnce(&[T]) -> R, plus an iterator iter that produces T items. You could combine them with f(&iter.collect::<Vec<_>>()), but this requires allocating memory for the Vec.

DebugWithInfcx

HashStableContext

Needed so we can use #[derive(HashStable_Generic)]

InferCtxtLike

Interner

Functions

debug_bound_var