rustc_type_ir
Module search_graph

Modules

global_cache 🔒

Structs

AvailableDepth 🔒

CycleHeads 🔒

All cycle heads a given goal depends on, ordered by their stack depth.

GlobalCache

NestedGoals 🔒

The nested goals of each stack entry and the path from the stack entry to that nested goal.

ProvisionalCacheEntry 🔒

A provisional result of an already computed goals which depends on other goals still on the stack.

SearchGraph

StackDepth

StackEntry 🔒

Stack entries of the evaluation stack. Its fields tend to be lazily when popping a child goal or completely immutable.

Enums

PathKind

In the initial iteration of a cycle, we do not yet have a provisional result. In the case we return an initial provisional result depending on the kind of cycle.

UsageKind

Traits

Cx

The search graph does not simply use Interner directly to enable its fuzzing without having to stub the rest of the interner. We don’t make this a super trait of Interner as users of the shared type library shouldn’t have to care about Input and Result as they are implementation details of the search graph.

Delegate