Enum rustc_type_ir::sty::TyKind

pub enum TyKind<I: Interner> {

    Bool,
    Char,
    Int(IntTy),
    Uint(UintTy),
    Float(FloatTy),
    Adt(I::AdtDef, I::GenericArgsRef),
    Foreign(I::DefId),
    Str,
    Array(I::Ty, I::Const),
    Slice(I::Ty),
    RawPtr(I::TypeAndMut),
    Ref(I::Region, I::Ty, I::Mutability),
    FnDef(I::DefId, I::GenericArgsRef),
    FnPtr(I::PolyFnSig),
    Dynamic(I::ListBinderExistentialPredicate, I::Region, DynKind),
    Closure(I::DefId, I::GenericArgsRef),
    Generator(I::DefId, I::GenericArgsRef, I::Movability),
    GeneratorWitness(I::BinderListTy),
    GeneratorWitnessMIR(I::DefId, I::GenericArgsRef),
    Never,
    Tuple(I::ListTy),
    Alias(AliasKind, I::AliasTy),
    Param(I::ParamTy),
    Bound(DebruijnIndex, I::BoundTy),
    Placeholder(I::PlaceholderType),
    Infer(I::InferTy),
    Error(I::ErrorGuaranteed),

}

Defines the kinds of types used by the type system.

Types written by the user start out as hir::TyKind and get converted to this representation using AstConv::ast_ty_to_ty.

Variants

Bool

The primitive boolean type. Written as bool.

Char

The primitive character type; holds a Unicode scalar value (a non-surrogate code point). Written as char.

Int(IntTy)

A primitive signed integer type. For example, i32.

Uint(UintTy)

A primitive unsigned integer type. For example, u32.

Float(FloatTy)

A primitive floating-point type. For example, f64.

Adt(I::AdtDef, I::GenericArgsRef)

Algebraic data types (ADT). For example: structures, enumerations and unions.

For example, the type List<i32> would be represented using the AdtDef for struct List<T> and the args [i32].

Note that generic parameters in fields only get lazily substituted by using something like adt_def.all_fields().map(|field| field.ty(tcx, args)).

Foreign(I::DefId)

An unsized FFI type that is opaque to Rust. Written as extern type T.

Str

The pointee of a string slice. Written as str.

Array(I::Ty, I::Const)

An array with the given length. Written as [T; N].

Slice(I::Ty)

The pointee of an array slice. Written as [T].

RawPtr(I::TypeAndMut)

A raw pointer. Written as *mut T or *const T

Ref(I::Region, I::Ty, I::Mutability)

A reference; a pointer with an associated lifetime. Written as &'a mut T or &'a T.

FnDef(I::DefId, I::GenericArgsRef)

The anonymous type of a function declaration/definition. Each function has a unique type.

For the function fn foo() -> i32 { 3 } this type would be shown to the user as fn() -> i32 {foo}.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar = foo; // bar: fn() -> i32 {foo}

FnPtr(I::PolyFnSig)

A pointer to a function. Written as fn() -> i32.

Note that both functions and closures start out as either FnDef or Closure which can be then be coerced to this variant.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar: fn() -> i32 = foo;

Dynamic(I::ListBinderExistentialPredicate, I::Region, DynKind)

A trait object. Written as dyn for<'b> Trait<'b, Assoc = u32> + Send + 'a.

Closure(I::DefId, I::GenericArgsRef)

The anonymous type of a closure. Used to represent the type of |a| a.

Closure args contain both the - potentially substituted - generic parameters of its parent and some synthetic parameters. See the documentation for ClosureArgs for more details.

Generator(I::DefId, I::GenericArgsRef, I::Movability)

The anonymous type of a generator. Used to represent the type of |a| yield a.

For more info about generator args, visit the documentation for GeneratorArgs.

GeneratorWitness(I::BinderListTy)

A type representing the types stored inside a generator. This should only appear as part of the GeneratorArgs.

Note that the captured variables for generators are stored separately using a tuple in the same way as for closures.

Unlike upvars, the witness can reference lifetimes from inside of the generator itself. To deal with them in the type of the generator, we convert them to higher ranked lifetimes bound by the witness itself.

Looking at the following example, the witness for this generator may end up as something like for<'a> [Vec<i32>, &'a Vec<i32>]:

#![feature(generators)]
|a| {
    let x = &vec![3];
    yield a;
    yield x[0];
}

GeneratorWitnessMIR(I::DefId, I::GenericArgsRef)

A type representing the types stored inside a generator. This should only appear as part of the GeneratorArgs.

Unlike upvars, the witness can reference lifetimes from inside of the generator itself. To deal with them in the type of the generator, we convert them to higher ranked lifetimes bound by the witness itself.

This variant is only using when drop_tracking_mir is set. This contains the DefId and the GenericArgsRef of the generator. The actual witness types are computed on MIR by the mir_generator_witnesses query.

Looking at the following example, the witness for this generator may end up as something like for<'a> [Vec<i32>, &'a Vec<i32>]:

#![feature(generators)]
|a| {
    let x = &vec![3];
    yield a;
    yield x[0];
}

Never

The never type !.

Tuple(I::ListTy)

A tuple type. For example, (i32, bool).

Alias(AliasKind, I::AliasTy)

A projection, opaque type, weak type alias, or inherent associated type. All of these types are represented as pairs of def-id and args, and can be normalized, so they are grouped conceptually.

Param(I::ParamTy)

A type parameter; for example, T in fn f<T>(x: T) {}.

Bound(DebruijnIndex, I::BoundTy)

Bound type variable, used to represent the 'a in for<'a> fn(&'a ()).

For canonical queries, we replace inference variables with bound variables, so e.g. when checking whether &'_ (): Trait<_> holds, we canonicalize that to for<'a, T> &'a (): Trait<T> and then convert the introduced bound variables back to inference variables in a new inference context when inside of the query.

It is conventional to render anonymous bound types like ^N or ^D_N, where N is the bound variable’s anonymous index into the binder, and D is the debruijn index, or totally omitted if the debruijn index is zero.

See the rustc-dev-guide for more details about higher-ranked trait bounds and canonical queries.

Placeholder(I::PlaceholderType)

A placeholder type, used during higher ranked subtyping to instantiate bound variables.

It is conventional to render anonymous placeholer types like !N or !U_N, where N is the placeholder variable’s anonymous index (which corresponds to the bound variable’s index from the binder from which it was instantiated), and U is the universe index in which it is instantiated, or totally omitted if the universe index is zero.

Infer(I::InferTy)

A type variable used during type checking.

Similar to placeholders, inference variables also live in a universe to correctly deal with higher ranked types. Though unlike placeholders, that universe is stored in the InferCtxt instead of directly inside of the type.

Error(I::ErrorGuaranteed)

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages.

Implementations

impl<I: Interner> TyKind<I>

pub fn is_primitive(&self) -> bool