Module rustc_type_ir::ty_kind

Re-exports

pub use self::closure::*;

Modules

closure 🔒

Structs

AliasTy

Represents the projection of an associated, opaque, or lazy-type-alias type.

FloatVid

A floating-point (f32 or f64) type variable ID.

FnSig

IntVid

An integral (u32, i32, usize, etc.) type variable ID.

TyVid

A type variable ID.

TypeAndMut

Enums

AliasTyKind

DynKind

Specifies how a trait object is represented.

FloatTy

FloatVarValue

InferTy

A placeholder for a type that hasn’t been inferred yet.

IntTy

IntVarValue

TyKind

Defines the kinds of types used by the type system.

UintTy

Functions

tykind_discriminant 🔒